+++ spikingjelly/activation_based/neuron/integrate_and_fire.py
@@ -303,7 +303,7 @@ class IFNode(BaseNode):
 
     def _inductor_multi_step_forward(self, x_seq: torch.Tensor):
         self.v_float_to_tensor(x_seq[0])
-        spike_seq, self.v, v_seq = functional.if_scan_inductor(
+        spike_seq, self.v, v_seq = functional.if_multi_step_inductor(
             x_seq,
             self.v,
             self.v_threshold,
@@ -1113,7 +1113,7 @@ class ActivationAwareIFNode(base.MemoryModule):
             channel_size = 1
             inner_size = x_seq[0].numel()
 
-        spike_seq, self.v, v_seq = functional.activation_aware_if_scan_triton(
+        spike_seq, self.v, v_seq = functional.activation_aware_if_multi_step_triton(
             x_seq,
             self.v,
             threshold,