+++ spikingjelly/activation_based/neuron/lif.py
@@ -591,7 +591,7 @@ class LIFNode(BaseNode):
 
     def _inductor_multi_step_forward(self, x_seq: torch.Tensor):
         self.v_float_to_tensor(x_seq[0])
-        spike_seq, self.v, v_seq = functional.lif_scan_inductor(
+        spike_seq, self.v, v_seq = functional.lif_multi_step_inductor(
             x_seq,
             self.v,
             self.tau,